Fixed window resizing. Added helix field.
Added functions 'setText' and 'setVectorField'.

Duplicated matFromVec and made a new version
(slightly changed) to have the x vector of the rotation matrix comply
with the incoming x dir. This is because the vtkArrowSource is pointed
down the positive x axis. One of these highly redundant functions will
be removed in the future.