BuildNoiseColumnMatrix files
The files for the BuildNoiseColumnMatrix module. So far pretty much
everything is working, except that when the algorithm is run it changes
the input. Output, however, is correctly calculated and sent to the
output.

Also, this was originally only a module (Not algorithm file), but I
changed it so that there is both an algorithm and module file.