Working version
This update is working. Successfully converts the matrix to a string.
Only possible issue is the module output because I cant see any modules
that take a string input, other than the write modules, but those want a
file name, not data. However, printing the string to the consol gives
the correct results.