Issue #1173: ViewScene Testing Mode (#1447)
Connected a ReportMatrixInfo to all of the ViewScene modules in the
regression networks with the exception of ViewSceneBackgroundColor
since there was no data to report and would result in an error. The
following networks finished with errors on the ReportMatrixInfo modules:
ClipFieldTorso
disabledModules
setuptdcs_colin27_patchelc

These errors will have to be fixed.